# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/travel/current": {
      "get": {
        "operationId": "TravelController_getCurrentTravel",
        "summary": "获取当前旅行",
        "description": "获取用户当前正在进行的旅行信息",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/TravelDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "没有正在进行的旅行",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "travel"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/travel/start": {
      "post": {
        "operationId": "TravelController_startTravel",
        "summary": "开始旅行",
        "description": "开始新的旅行（单人/双人）",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartTravelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/TravelDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误或已有正在进行的旅行",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "城市不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "travel"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/travel/switch-to-dual": {
      "post": {
        "operationId": "TravelController_switchToDual",
        "summary": "单人旅行切换为双人",
        "description": "将当前单人旅行切换为双人旅行",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchToDualDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/TravelDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误或已经是双人旅行",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "没有正在进行的旅行",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "travel"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/travel/companions": {
      "get": {
        "operationId": "TravelController_getCurrentCompanions",
        "summary": "获取当前旅伴列表",
        "description": "获取当前旅行中的旅伴信息（仅双人旅行有）",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/TravelCompanionsResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "travel"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/travel/sync": {
      "post": {
        "operationId": "TravelController_syncSteps",
        "summary": "同步步数",
        "description": "从前端设备同步步数到服务器",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncStepsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "boolean"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误或没有正在进行的旅行",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "travel"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/travel/statistics": {
      "get": {
        "operationId": "TravelController_getStatistics",
        "summary": "获取旅行统计",
        "description": "获取用户的历史旅行统计信息（整体统计：旅行城市数、所有步数、消耗卡路里）",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/TravelStatisticsDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "travel"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/travel/statistics/cities": {
      "get": {
        "operationId": "TravelController_getCityTravelStatistics",
        "summary": "获取按城市分组的旅行统计",
        "description": "获取用户按城市分组的详细旅行统计信息（每个城市的旅行天数、所有步数、消耗卡路里）",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/CityTravelStatsListResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "travel"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/travel/invitation/generate": {
      "post": {
        "operationId": "TravelController_generateInvitation",
        "summary": "生成邀请码",
        "description": "生成一个邀请码，用于邀请其他用户成为旅伴",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/GenerateInvitationResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "travel"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/travel/invitation/validate": {
      "post": {
        "operationId": "TravelController_validateInvitation",
        "summary": "验证邀请码",
        "description": "验证邀请码是否有效，并返回邀请者信息",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateInvitationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ValidateInvitationSuccessResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "travel"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/travel/invitation/accept": {
      "post": {
        "operationId": "TravelController_acceptInvitation",
        "summary": "接受邀请",
        "description": "接受邀请码，建立旅伴关系",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AcceptInvitationResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "travel"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/travel/partnerships": {
      "get": {
        "operationId": "TravelController_getPartnerships",
        "summary": "获取旅伴关系列表",
        "description": "获取用户的所有旅伴关系",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PartnershipsResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "travel"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/travel/partnerships/{id}": {
      "delete": {
        "operationId": "TravelController_unbindPartnership",
        "summary": "解除旅伴关系",
        "description": "解除旅伴关系（需24小时确认）",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UnbindResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误或有正在进行的旅行",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "旅伴关系不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "travel"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/travel/partnerships/cancel-unbind": {
      "post": {
        "operationId": "TravelController_cancelUnbind",
        "summary": "取消解除请求",
        "description": "取消解除旅伴关系的请求",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelUnbindDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/CancelUnbindResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "旅伴关系不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "travel"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Claw Step API",
    "description": "Claw Step 应用后端 API 文档",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "ResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "状态码",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "提示信息",
            "example": "success"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          }
        },
        "required": [
          "code",
          "message",
          "data"
        ]
      },
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "状态码",
            "example": 400
          },
          "message": {
            "type": "string",
            "description": "错误信息",
            "example": "请求参数错误"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "TravelDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "旅行ID",
            "example": "uuid-string"
          },
          "userId": {
            "type": "string",
            "description": "用户ID",
            "example": "uuid-string"
          },
          "cityId": {
            "type": "string",
            "description": "城市ID",
            "example": "uuid-string"
          },
          "type": {
            "type": "string",
            "description": "旅行类型",
            "example": "single",
            "enum": [
              "single",
              "dual"
            ]
          },
          "partnerId": {
            "type": "string",
            "description": "旅伴ID",
            "example": "uuid-string"
          },
          "partnershipId": {
            "type": "string",
            "description": "旅伴关系ID",
            "example": "uuid-string"
          },
          "status": {
            "type": "string",
            "description": "旅行状态",
            "example": "active",
            "enum": [
              "active",
              "completed",
              "paused"
            ]
          },
          "startDate": {
            "format": "date-time",
            "type": "string",
            "description": "开始时间",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "endDate": {
            "format": "date-time",
            "type": "string",
            "description": "结束时间",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "completedAt": {
            "format": "date-time",
            "type": "string",
            "description": "完成时间",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "totalSteps": {
            "type": "number",
            "description": "总步数",
            "example": 5000
          },
          "totalCalories": {
            "type": "number",
            "description": "总卡路里",
            "example": 200
          },
          "days": {
            "type": "number",
            "description": "已旅行天数",
            "example": 3
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "更新时间",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "city": {
            "description": "城市信息",
            "allOf": [
              {
                "$ref": "#/components/schemas/CityVo"
              }
            ]
          },
          "partner": {
            "description": "旅伴信息",
            "allOf": [
              {
                "$ref": "#/components/schemas/PartnerVo"
              }
            ]
          }
        },
        "required": [
          "id",
          "userId",
          "cityId",
          "type",
          "status",
          "startDate",
          "totalSteps",
          "totalCalories",
          "days",
          "createdAt",
          "updatedAt"
        ]
      },
      "StartTravelDto": {
        "type": "object",
        "properties": {
          "cityId": {
            "type": "string",
            "description": "城市ID",
            "example": "uuid-string"
          },
          "type": {
            "type": "string",
            "description": "旅行类型",
            "enum": [
              "single",
              "dual"
            ],
            "example": "single"
          },
          "partnerId": {
            "type": "string",
            "description": "旅伴ID（双人旅行时必需）",
            "example": "uuid-string"
          },
          "partnershipId": {
            "type": "string",
            "description": "旅伴关系ID（双人旅行时必需）",
            "example": "uuid-string"
          }
        },
        "required": [
          "cityId",
          "type"
        ]
      },
      "SwitchToDualDto": {
        "type": "object",
        "properties": {
          "partnerId": {
            "type": "string",
            "description": "旅伴ID",
            "example": "uuid-string"
          },
          "partnershipId": {
            "type": "string",
            "description": "旅伴关系ID",
            "example": "uuid-string"
          }
        },
        "required": [
          "partnerId",
          "partnershipId"
        ]
      },
...
      "TravelCompanionsResponseDto": {
        "type": "object",
        "properties": {
          "companions": {
            "description": "旅伴列表",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TravelCompanionDto"
            }
          }
        },
        "required": [
          "companions"
        ]
      },
      "SyncStepsDto": {
        "type": "object",
        "properties": {
          "steps": {
            "type": "number",
            "description": "步数",
            "example": 3500
          },
          "calories": {
            "type": "number",
            "description": "卡路里",
            "example": 120
          },
          "date": {
            "type": "string",
            "description": "日期（ISO格式），默认今天",
            "example": "2024-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "steps"
        ]
      },
      "TravelStatisticsDto": {
        "type": "object",
        "properties": {
          "totalCities": {
            "type": "number",
            "description": "旅行城市数",
            "example": 2
          },
          "totalSteps": {
            "type": "number",
            "description": "所有步数",
            "example": 43558
          },
          "totalCalories": {
            "type": "number",
            "description": "消耗卡路里",
            "example": 5848
          }
        },
        "required": [
          "totalCities",
          "totalSteps",
          "totalCalories"
        ]
      },
...
      "CityTravelStatsListResponseDto": {
        "type": "object",
        "properties": {
          "cities": {
            "description": "城市统计列表",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CityTravelStatsDto"
            }
          }
        },
        "required": [
          "cities"
        ]
      },
      "GenerateInvitationResponseDto": {
        "type": "object",
        "properties": {
          "invitationCode": {
            "type": "string",
            "description": "邀请码",
            "example": "ABC12345"
          },
          "expiresAt": {
            "type": "string",
            "description": "过期时间",
            "example": "2024-01-08T00:00:00.000Z"
          },
          "qrCodeUrl": {
            "type": "string",
            "description": "二维码URL",
            "example": "https://example.com/qr/ABC12345"
          }
        },
        "required": [
          "invitationCode",
          "expiresAt"
        ]
      },
      "ValidateInvitationSuccessResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "是否成功",
            "example": true
          },
          "data": {
            "type": "object",
            "description": "邀请信息"
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
      "ValidateInvitationDto": {
        "type": "object",
        "properties": {
          "invitationCode": {
            "type": "string",
            "description": "邀请码",
            "example": "ABC12345"
          }
        },
        "required": [
          "invitationCode"
        ]
      },
      "AcceptInvitationResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "是否成功",
            "example": true
          },
          "data": {
            "type": "object",
            "description": "旅伴关系信息"
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
      "AcceptInvitationDto": {
        "type": "object",
        "properties": {
          "invitationCode": {
            "type": "string",
            "description": "邀请码",
            "example": "ABC12345"
          }
        },
        "required": [
          "invitationCode"
        ]
      },
...
      "CancelUnbindDto": {
        "type": "object",
        "properties": {
          "partnershipId": {
            "type": "string",
            "description": "解除请求ID",
            "example": "uuid-string"
          }
        },
        "required": [
          "partnershipId"
        ]
      },
...
    }
  }
}